Renames 'writer' module to 'ion_writer'
