+++ src/binary/binary_writer.rs
@@ -1,9 +1,9 @@
 use crate::binary::raw_binary_writer::{RawBinaryWriter, RawBinaryWriterBuilder};
 use crate::constants::v1_0::system_symbol_ids;
+use crate::ion_writer::IonWriter;
 use crate::raw_symbol_token_ref::{AsRawSymbolTokenRef, RawSymbolTokenRef};
 use crate::result::{illegal_operation, IonResult};
 use crate::types::{Decimal, Int, IonType, SymbolId, Timestamp};
-use crate::writer::IonWriter;
 use crate::SymbolTable;
 use delegate::delegate;
 use std::io::Write;
