+++ src/element/element_stream_writer.rs
@@ -1,9 +1,9 @@
 use crate::element::builders::StructBuilder;
 use crate::element::{Annotations, Element, IntoAnnotatedElement, Value};
+use crate::ion_writer::IonWriter;
 use crate::raw_symbol_token_ref::AsRawSymbolTokenRef;
 use crate::result::illegal_operation;
 use crate::types::Bytes;
-use crate::writer::IonWriter;
 use crate::{
     Decimal, Int, IonResult, IonType, RawSymbolTokenRef, Str, Symbol, SymbolTable, Timestamp,
 };
@@ -299,8 +299,8 @@ mod tests {
     use crate::element::builders::{SequenceBuilder, StructBuilder};
     use crate::result::IonResult;
 
+    use crate::ion_writer::IonWriter;
     use crate::types::{Bytes, Timestamp};
-    use crate::writer::IonWriter;
     use crate::{Decimal, IonType, Symbol};
 
     #[track_caller]
