+++ src/element/mod.rs
@@ -16,9 +16,9 @@ use crate::binary::binary_writer::BinaryWriterBuilder;
 use crate::element::builders::{SequenceBuilder, StructBuilder};
 use crate::element::reader::ElementReader;
 use crate::ion_data::{IonEq, IonOrd};
+use crate::ion_writer::IonWriter;
 use crate::text::text_formatter::IonValueFormatter;
 use crate::text::text_writer::TextWriterBuilder;
-use crate::writer::IonWriter;
 use crate::{ion_data, Decimal, Int, IonResult, IonType, Str, Symbol, Timestamp};
 use std::cmp::Ordering;
 use std::fmt::{Display, Formatter};
