+++ src/text/text_writer.rs
@@ -1,10 +1,10 @@
 use crate::element::writer::TextKind;
+use crate::ion_writer::IonWriter;
 use crate::raw_symbol_token_ref::{AsRawSymbolTokenRef, RawSymbolTokenRef};
 use crate::result::IonResult;
 use crate::text::raw_text_writer::RawTextWriter;
 use crate::text::raw_text_writer::RawTextWriterBuilder;
 use crate::types::{Decimal, Timestamp};
-use crate::writer::IonWriter;
 use crate::{Int, IonType, SymbolTable};
 use delegate::delegate;
 use std::io::Write;
